🐛 fix(knowledge-base): enforce restrictions on the OpenAPI surface and batch the lock-badge flag
- move the knowledgeBaseAccess helpers to a server service so packages/openapi
  can reuse them (router _helpers re-export unchanged)
- OpenAPI KB list/detail and file list/detail/url/parse/chunk-status/upload
  now honor the restricted-KB policy
- getKnowledgeBases returns a memberRestricted flag so the sidebar lock badge
  no longer issues one permission request per row
