+++ src/features/ResourceHome/Layout/Body/LibraryList/index.tsx
@@ -84,6 +84,7 @@ const LibraryList = memo(() => {
             description={item.description}
             id={item.id}
             key={item.id}
+            memberRestricted={(item as { memberRestricted?: boolean }).memberRestricted}
             name={item.name}
             userId={item.userId}
             visibility={item.visibility}
